+++ app/src/main/java/tom/r/projetandroid4a/Controller/Fragment/FragmentWakfu.java
@@ -1,4 +1,4 @@
-package tom.r.projetandroid4a.Fragment;
+package tom.r.projetandroid4a.Controller.Fragment;
 
 import android.os.Bundle;
 import android.support.annotation.NonNull;
@@ -19,21 +19,21 @@ import tom.r.projetandroid4a.R;
 /**
  * @author Tom
  */
-public class FragmentDofus extends Fragment {
+public class FragmentWakfu extends Fragment{
     public View view;
 
-    public FragmentDofus(){
+    public FragmentWakfu(){
     }
 
     @Nullable
     @Override
     public View onCreateView(@NonNull LayoutInflater inflater, @Nullable ViewGroup container, @Nullable Bundle savedInstanceState) {
-        view=inflater.inflate(R.layout.fragment_dofus, container, false);
+        view=inflater.inflate(R.layout.fragment_wakfu, container, false);
 
         // Reception des données
         String id = getActivity().getIntent().getExtras().getString("p_id");
         String name = getActivity().getIntent().getExtras().getString("p_name");
-        String description_dofus = getActivity().getIntent().getExtras().getString("p_description_dofus");
+        String description_wakfu = getActivity().getIntent().getExtras().getString("p_description_wakfu");
         String image = getActivity().getIntent().getExtras().getString("p_image");
         String classe = getActivity().getIntent().getExtras().getString("p_classe");
         String gender = getActivity().getIntent().getExtras().getString("p_gender");
@@ -44,8 +44,8 @@ public class FragmentDofus extends Fragment {
         String image_sort1 = getActivity().getIntent().getExtras().getString("p_image_sort1");
         String image_sort2 = getActivity().getIntent().getExtras().getString("p_image_sort2");
         String image_sort3 = getActivity().getIntent().getExtras().getString("p_image_sort3");
-        String image_dofus = getActivity().getIntent().getExtras().getString("p_image_dofus");
-        String lien_dofus = getActivity().getIntent().getExtras().getString("p_lien_dofus");
+        String image_wakfu = getActivity().getIntent().getExtras().getString("p_image_wakfu");
+        String lien_wakfu = getActivity().getIntent().getExtras().getString("p_lien_wakfu");
 
         // CollapsingToolbar
         CollapsingToolbarLayout collapsingToolbarLayout = view.findViewById(R.id.collapsingtoolbar_id);
@@ -55,7 +55,7 @@ public class FragmentDofus extends Fragment {
         TextView pId = view.findViewById(R.id.p_id);
         TextView pName = view.findViewById(R.id.p_name);
         TextView pClasse = view.findViewById(R.id.p_classe);
-        TextView pDescription_dofus = view.findViewById(R.id.p_description_dofus);
+        TextView pDescription_wakfu = view.findViewById(R.id.p_description_wakfu);
         TextView pGender = view.findViewById(R.id.p_gender);
         TextView pSort1 = view.findViewById(R.id.p_sort1);
         TextView pSort2 = view.findViewById(R.id.p_sort2);
@@ -63,28 +63,27 @@ public class FragmentDofus extends Fragment {
         ImageView pImageSort1 = view.findViewById(R.id.p_image_sort1);
         ImageView pImageSort2 = view.findViewById(R.id.p_image_sort2);
         ImageView pImageSort3 = view.findViewById(R.id.p_image_sort3);
-        ImageView pImage = view.findViewById(R.id.p_image_dofus);
-        TextView pLien = view.findViewById(R.id.p_lien_dofus);
+        ImageView pImage = view.findViewById(R.id.p_image_wakfu);
+        TextView pLien = view.findViewById(R.id.p_lien_wakfu);
         ImageView pIcone = view.findViewById(R.id.p_icone);
-
 
         // setting des valeurs de chaque view
         pName.setText(name);
         pId.setText(id);
         pClasse.setText(classe);
-        pDescription_dofus.setText(description_dofus);
+        pDescription_wakfu.setText(description_wakfu);
         pGender.setText(gender);
         pSort1.setText(sort1);
         pSort2.setText(sort2);
         pSort3.setText(sort3);
-        pLien.setText(lien_dofus);
+        pLien.setText(lien_wakfu);
 
         // requêtes option pour les images avec Glide en fonction du fond de l'ImageView
         RequestOptions requestOptions = new RequestOptions().centerCrop().placeholder(R.drawable.loading_shape).error(R.drawable.loading_shape);
         RequestOptions requestOptions2 = new RequestOptions().centerCrop().placeholder(R.drawable.bgg).error(R.drawable.bgg);
 
         // chargement des images avec Glide
-        Glide.with(this).load(image_dofus).apply(requestOptions).into(pImage);
+        Glide.with(this).load(image_wakfu).apply(requestOptions).into(pImage);
         Glide.with(this).load(icone).apply(requestOptions2).into(pIcone);
         Glide.with(this).load(image_sort1).apply(requestOptions2).into(pImageSort1);
         Glide.with(this).load(image_sort2).apply(requestOptions2).into(pImageSort2);
